% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/county_state_fips_table.R
\docType{data}
\name{county_state_fips}
\alias{county_state_fips}
\title{county_state_fips}
\format{A data.frame with 6 variables indicating state abbr., state name, county name,
        state fips, county fips, and total fips}
\usage{
data(county_state_fips)
}
\description{
county_state_fips
}
\references{
\href{www.mdreducation.com/pdfs/US_FIPS_Codes.xls}
}
\keyword{fips}
