% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_R_data_transfer.R
\name{db_type_converter}
\alias{db_type_converter}
\title{db_type_converter}
\usage{
db_type_converter(dbname, data)
}
\arguments{
\item{dbname}{The name of the database. Used to distinguish data.}

\item{data}{The actual data.frame to convert.}
}
\value{
The modified data.frame
}
\description{
This function converts the type to align with the requirement. See requirement online.
}
\examples{
# data is the output from any get_from_db function
data <- get_from_db_usr("SELECT loadid FROM hedonics_new.sd_hedonics_new")
# convert the type of the columns
data <- db_type_converter("zillow_2017_nov", data)
}
