% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/googlekml.R
\name{get_cairo_usr}
\alias{get_cairo_usr}
\title{get_cairo_usr}
\usage{
get_cairo_usr(query, database_name = "googlekml", host_ip = "3.90.230.196")
}
\arguments{
\item{query}{A string specifying the query sent to database}

\item{database_name}{A string indicating the database name}

\item{host_ip}{A string indicating the ip address of the database VM}
}
\value{
A data.frame returned by the given query.
}
\description{
This function gets from database according to a user-specified query.
}
\examples{
# Select single field from a given table
data <- get_from_db_usr("SELECT COUNT(*) FROM routes")
 
See descriptions of get_from_db_usr for more details on possible queries
}
