% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_R_data_transfer.R
\name{get_from_db_fips}
\alias{get_from_db_fips}
\title{get_from_db_fips}
\usage{
get_from_db_fips(fips, columns = "*", database_name = "zillow_2017_nov",
  host_ip = "141.142.209.139")
}
\arguments{
\item{fips}{A vector of fips}

\item{columns}{A vector of column names to export. Default to all columns (i.e. "*").}

\item{database_name}{A string indicating the database name}

\item{host_ip}{A string indicating the ip address of the database VM}
}
\value{
A data.frame including all data from the given fips.
}
\description{
This function gets a data.frame including all data from the given fips. Use previous
             function.
}
