% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_R_data_transfer.R
\name{get_from_db_usr}
\alias{get_from_db_usr}
\title{get_from_db_usr}
\usage{
get_from_db_usr(query, database_name = "zillow_2017_nov",
  host_ip = "141.142.209.139")
}
\arguments{
\item{query}{A string specifying the query sent to database}

\item{database_name}{A string indicating the database name}

\item{host_ip}{A string indicating the ip address of the database VM}
}
\value{
A data.frame returned by the given query.
}
\description{
This function gets from database according to a user-specified query.
}
\examples{
data <- get_from_db_usr("SELECT loadid FROM hedonics_new.sd_hedonics_new")
}
