% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infousa.R
\name{get_infousa_fid}
\alias{get_infousa_fid}
\title{get_infousa_fid}
\usage{
get_infousa_fid(startyear, endyear, fid, columns = "*")
}
\arguments{
\item{startyear}{The first year to get data}

\item{endyear}{The last year to get data}

\item{fid}{A vector of characters indicating familyids to get data from.}

\item{columns}{A vector of column names to export. Default to all columns (i.e. "*").}
}
\value{
A data.frame
}
\description{
This function gets a data.frame including all data from the given vector of familyid from given years.
}
\examples{
 test <- get_infousa_fid(2006, 2007, c("54299524", "54320129"))
}
