% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infousa.R
\name{get_infousa_multiyear}
\alias{get_infousa_multiyear}
\title{get_infousa_multiyear}
\usage{
get_infousa_multiyear(loc, startyear, endyear, tract = "*", columns = "*",
  method = "fips")
}
\arguments{
\item{loc}{A vector of integers indicating fips ("fips" method) or a data.frame with first 2
columns being state abbr. & county name ("names" method).
If the entire state is needed, set county code to 000 ("fips") or set county name to "*" ("names").}

\item{startyear}{The first year to get data}

\item{endyear}{The last year to get data}

\item{tract}{A vector of integers or chars indicating tract in the county.
Note that tracts are unique only in the current county. Default to all tracts.}

\item{columns}{A vector of column names to export. Default to all columns (i.e. "*").}

\item{method}{Method for input. Choose between "fips" and "names". Default to "fips".}
}
\value{
A data.frame including data from all years, fips and tract
}
\description{
This function gets a data.frame including all data from the given location from all years.
}
\examples{
 test <- get_infousa_location("01001")
 test <- get_infousa_location("02020", tract=c(001802,002900))
}
