% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infousa.R
\name{get_infousa_state_city}
\alias{get_infousa_state_city}
\title{get_from_db_state_city}
\usage{
get_infousa_state_city(state_city, year, columns = "*", append = TRUE)
}
\arguments{
\item{state_city}{A data.frame with two columns representing state and county name}

\item{columns}{A vector of column names to export. Default to all columns (i.e. "*").}

\item{append}{If append is true, return a single data.frame with rows appended, otherwise a
list of data.frames from each state.}
}
\value{
A data.frame including all data from the given state and county
}
\description{
This function gets a data.frame including all data from the given state abbr. and county.
}
\examples{
table <- data.frame(state=c("IL"),city=c("Chicago"))
data <- get_infousa_state_city(table)
}
