% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infousa.R
\name{get_infousa_zip}
\alias{get_infousa_zip}
\title{get_infousa_zip}
\usage{
get_infousa_zip(startyear, endyear, zip, columns = "*")
}
\arguments{
\item{startyear}{The first year to get data}

\item{endyear}{The last year to get data}

\item{zip}{A vector of characters indicating zipcodes to get data from.}

\item{columns}{A vector of column names to export. Default to all columns (i.e. "*").}
}
\value{
A data.frame including data from all years, fips and tract
}
\description{
This function gets a data.frame including all data from the given location from all years.
}
\examples{
 test <- get_infousa_zip(2016, 2017, c("61801", "61820"))
}
