% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zip_state_city.R
\name{get_state_city_zipcode}
\alias{get_state_city_zipcode}
\title{get_state_city_zipcode}
\usage{
get_state_city_zipcode(zips)
}
\arguments{
\item{zips}{A vector of zip codes}
}
\value{
A data.frame with first column as state, second column as zipcode,
        and last column the corresponding city.
}
\description{
This function convert a vector of zipcodes to corresponding state-city pairs
}
