% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_R_data_transfer.R
\name{get_state_county}
\alias{get_state_county}
\title{get_state_county}
\usage{
get_state_county(fips)
}
\arguments{
\item{fips}{A vector of fips numbers stored as TEXT (characters)}
}
\value{
A data.frame with first column as state abbreviation and second column as county name,
        all in lower case.
}
\description{
This function convert a vector of fips to corresponding state-county pairs
}
