% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/county_state_fips_table.R
\name{get_state_county_by_fips}
\alias{get_state_county_by_fips}
\title{get_state_county_by_fips}
\usage{
get_state_county_by_fips(fips)
}
\arguments{
\item{fips}{A vector of fips numbers}
}
\value{
A data.frame with first column as state abbreviation, second column as county name,
        third column county_code, and last column fips.
}
\description{
This function convert a vector of fips to corresponding state-county pairs
}
