% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_points.R
\docType{data}
\name{split_points}
\alias{split_points}
\title{split_points}
\format{A integer array that contains the split points of familyid}
\usage{
data(split_points)
}
\description{
split_points
}
\keyword{familyid}
