% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zip_state_city.R
\docType{data}
\name{zip_state_city}
\alias{zip_state_city}
\title{zip_state_city}
\format{A data.frame with combinations of zipcode, city and state names}
\usage{
data(zip_state_city)
}
\description{
zip_state_city
}
\references{
\url{www.pier2pier.com/links/files/Countrystate/USA-Zip.xls}
}
\keyword{city,}
\keyword{state}
\keyword{zip,}
\keyword{zipcode,}
