% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_R_data_transfer.R
\name{get_from_db}
\alias{get_from_db}
\title{get_from_db}
\usage{
get_from_db(spec, arg, columns = "*", database_name = "zillow_2017_nov",
  host_ip = "141.142.209.139", user = FALSE)
}
\arguments{
\item{spec}{A single string indicating which column (field) to specify.}

\item{arg}{A dataframe indicating how to specify.}

\item{columns}{A vector of column names to export. Default to all columns (i.e. "*").}

\item{database_name}{A string indicating the database name}

\item{host_ip}{A string indicating the ip address of the database VM}

\item{user}{If TRUE, the function regards arg as the given query sent to database.
Note that user-defined query should be used whenever two specifications
are present.}
}
\value{
A data.frame including all data from the given specs.
}
\description{
This function is a wrapper for all get_from_db method. By specifying its arguments,
             it calls different functions in the script. The specification can be either previously
             recorded fields, or user-defined queries. It returns a data.frame including all data
             specified.
}
