% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infousa.R
\name{get_infousa_location}
\alias{get_infousa_location}
\title{get_infousa_location}
\usage{
get_infousa_location(single_year, fips, tract = "*", columns = "*",
  append = TRUE)
}
\arguments{
\item{fips}{A vector of integers indicating fips.
If the entire state is wanted, set county code to 000.}

\item{tract}{A vector of integers or chars indicating tract in the county.
Note that tracts are unique only in the current county. Default to all tracts.}

\item{columns}{A vector of column names to export. Default to all columns (i.e. "*").}

\item{append}{If append is true, return a single data.frame with rows appended, otherwise a
list of data.frames from each state.}

\item{year}{An integer indicating year}
}
\value{
A data.frame including all data from the given year, fips and tract
}
\description{
This function gets a data.frame including all data from the given state abbr. and county.
}
\examples{
 test <- get_infousa_location(2006, 01001)
 test <- get_infousa_location(2006, 02020, tract=c(001802,002900))
}
